refactor(html-reporter): extract scenario matching into own module
Move parseScenarioParameters, findMatchingScenario, filterByTags,
and browserName into scenarioMatchingStrategies.ts. Reduces total
complexity of useScenarioDetail.ts, which now focuses solely on
hook state management and view data derivation.
